ci: ASAN build with DLUAJIT_ENABLE_GC64=ON
This patch is a pre-patch (see [1]) to add ASAN instrumentation to
the LuaJIT allocator. ASAN requires GC64 to be enabled because the
ASAN instrumentation of LuaJIT allocator is implemented for GC64
due to some difficulties building LuaJIT with `-m32` (see [1]).

Note: Until [1] is not completed, Tarantool with ASAN must be built
with `-DLUAJIT_USE_SYSMALLOC=ON'.

[1]: Issue #10231

NO_DOC=ci
NO_CHANGELOG=ci
NO_TEST=ci