schema: error msg for array instead of map/record
This commit addresses an issue where if a map/record field is
expected in a config field, and an array is received instead,
the error message thrown is not very detailed or user-friendly.

Also added a check for <schema object>:merge(a, b) to ensure
that if <schema object>.type = map or record, a and b must not
be arrays.
Previously:
```
local scalar = schema.scalar({type = 'string'})
local myschema = schema.new('myschema', schema.map({
    key = scalar,
    value = scalar,
}))
myschema = myschema:merge(myschema, {1, 2, 3}) -- ok
-- myschema:
--
-- 1: 1
-- 2: 2
-- methods: []
-- schema:
--   type: map
-- ...
```
Now: `myschema = myschema:merge(myschema, {1, 2, 3}) -- error`

Now `{[1] = true}` (see test/config-luatest/schema_test.lua:
test_validate_map) is defined as an array rather than a map with
key 1, since this map representation corresponds to the array
representation `{true}`.
So it was replaced with `{[2] = true}`, which is not an array.

Closes #10241
Closes #10242

NO_DOC=internal
NO_CHANGELOG=internal