box: grant/revoke user rights after upgrade
User/role privileges weren't granted/revoked after upgrading schema
based on an old snapshot from 2.11 to 3.x. Now the credentials applier
waits for the schema to be updated to the latest version and then
automatically starts managing configured privileges.

Fix tarantool/tarantool@8f43d0b is no longer relatable. So remove it.

Add error handling in credenetials applier for user to be notified if
granting/revoking process ended up in an error.

Still the solution isn't perfect because there may be some problems in
case of downgrading/upgrading in a single instance with the config
changes. It is so because of the upgraded status is waited only during
the first config apply of the instance run. The way to fix it and the
details are described at #10411.

NO_DOC=bugfix

Closes #9849