engine: consider raft and limbo states system
This commit makes force_recovery not to be applied to the recovery of
raft and limbo states, since allowing recovery with corrupted raft and
limbo states may lead to split brain.

These rows are now close to the system data in snapshot, so it's finally
easy to do.

Follow up tarantool/tarantool-ee#741

NO_DOC=internal
NO_CHANGELOG=internal