swim: fix crash on Tarantool shutdown
On Tarantool shutdown first swim worker fiber is finished as managed on
shutdown one. So later when swim is collected we got crash in swim_gc()
because `event_handler` is NULL at this point. There is no worker fiber
so we need to free swim synchronously.

The issue was hidden because we do not call lua_close() earlier. It had
little change to occur. Now it is revealed.

Closes #10495
Part-of #10211

NO_TEST=covered by existing tests
NO_DOC=bugfix