box: disallow write to space while it is being altered
Currently, we can write to space right after DDL is prepared and it
breaks MVCC - it aborts all in-progress transactions on prepare and
deletes outdated stories on commit so there is a window when
transactions can actually read outdated stories and it breaks some
internal invariants, Tarantool crashes as a result. Let's simply forbid
to write to space while DDL is not committed - it's much simplier than
invalidation of all stoies on prepare because we may need them when
commit fails.