test: fix gh-5430 test
The test blocked on the `wait_log` for registering replica2 on the master
node. The `wait_log` does not throw an error on timeout, and its return
value is ignored in the test, so the test simply continued after the
timeout and succeeded.

The reason the test blocked was that after 70a6883 we are unable to
register an anonymous replica while its relay has not stopped:
https://github.com/tarantool/tarantool/blob/9e616ab6f2f94ba2bcad3044187c474a1c05a8c5/src/box/box.cc#L4422-L4427

At the same time, the anonymous replica's relay cannot finish, because it
is blocked by `cbus_unpair` in `wal_clear_watcher`:
https://github.com/tarantool/tarantool/blob/9e616ab6f2f94ba2bcad3044187c474a1c05a8c5/src/box/relay.cc#L1068-L1074
https://github.com/tarantool/tarantool/blob/9e616ab6f2f94ba2bcad3044187c474a1c05a8c5/src/lib/core/cbus.h#L440-L442

To fix this let's move `ERRINJ_REPLICA_JOIN_DELAY` before
`box_register_replica`. This error injection is not used right now, so we
can reuse it here instead of `ERRINJ_WAL_DELAY`. This way the WAL won't get
blocked, and we will be able to get 2 concurrent replica registrations.

Let's also echo the `wait_log` result to make sure the fails in the future.

Closes #9617

NO_CHANGELOG=<test fix>
NO_DOC=<test fix>