connpool: make call() wait for any connection
The connection pool `call()` function used to connect to all of the
instances matching some specified static requiriments and synchroniouly
select one of the active. It was slower than possible because only
one active instance was needed to perform the call.

This patch makes the connection pool `call()` function work a bit
smarter. Now if the priorities (`prefer_ro`/`prefer_rw`
call modes) aren't used within the `connpool.call()` the pool will wait
for any connection matching the requirements as a candidate to perform
the call.

In case of using the priorities all of the candidates matching static
properties are fetched. It may be optimized using the fact that
technically it's possible to determine whether the reached replica has
the maiximum priority and can be used without waiting other ones.

@TarantoolBot document
Title: Experimental.connpool: new `filter()`, `call()` mode names

The `connpool.filter()` options got new modes:
- `alive` can be used to filter out only active connections matching
  other requirements.
- The default mode is now named `any` (used by default when the mode is
  nil). It can be used to ignore instance status completely.
- Other modes remained the same.

The `connpool.call()` default filtering mode has now got a name `alive`
(the mode is used if the mode is `nil`). It means "any alive instance".
Other candidate filters for call execution remained the same.

The `connpool.call()` function will now wait for any available instance
matching the requirements in case of the `alive/ro/rw` modes. And it
will try to connect to all of the matching instances if the mode is
`prefer_ro`/`prefer_rw`. So if the modes are one of the `prefer_*` the
function works slower in case there is at least one unavailable
instance.

Closes #10596 and #10598