box: unify field_map position
Now in different parts of code field map is can be pointed both
as to the first byte or to the next byte after last. Also it can
be pointed as uint32_t * or as char *. That looks misleading and
sometime not optimal.

This patch fixes the approach: field map is pointed by the next
after the last char in it.

Part of #10429

NO_DOC=internal
NO_TEST=covered by existing tests
NO_CHANGELOG=internal