box: store field map format id
If the field map is not empty, store field map format id with the
offsets. That would allow to use field map even for tuples in read
view.

Note that before this comming tuple data (which is the same as field
map pointer) was always equal to 2 by modulo 4. Now if there's a
non-empty field map, it becomes equal to 0 by module 4. This feature
is used for the sake of performance.

Part of #10429