txn: add 'wait' option to box.atomic
Similar to the raw `box.commit`, users can now specify `wait=...`
in `box.atomic`. The latter returns instantly if the wait mode was
not `complete` and the WAL queue isn't full.

The only non-trivial change is that `box.commit` in the
implementation of `box.atomic` had to be wrapped into raw `pcall`
instead of `call_at`, because if it fails with bad options, the
manual `box.rollback` is needed right after.

NO_DOC=later
NO_CHANGELOG=later

Part of #67