memtx: allow to abort checkpoint if it wasn't started
When checkpoint fails, we abort it in all engines even if it wasn't
started successfully. If it fails right from the start so that checkpoint
in memtx wasn't started, assertion in `memtx_engine_abort_checkpoint`
fails - memtx doesn't expect that checkpoint will be aborted if it
failed to start. Let's do the same thing as vinyl does - no-op if
there is no checkpoint in progress.

Closes #10265

NO_CHANGELOG=reproducible only with error injection
NO_DOC=bugfix