uri: fix parsing uppercased IPv6 addresses
Before this patch `uri.parse(<uri-string-with-ipv6-addres>)`
did not work correctly. In particular, it did not parse an
IPv6 address if it contained `A-F`. It is a regression caused
by commit 1376aad9b3fd ("Refactor src/uri.rl to support RFC3986
and add Lua bindings").

This patch fixes a bug where characters `A-F` are not
supported in IPv6.

Part of #9556

NO_DOC=bugfix