box: introduce the create_iterator_with_offset method
This method is used in order to perform the iterator creation with
the offset specified. The generic implementation performs the regular
iterator creation and skips the specified amount of tuples checking
the fiber slice in the way, but now it's possible to implement an
accelerated version of select with offset if an index supports that.

Part of #8204

NO_DOC=internal
NO_TEST=covered by existing tests
NO_CHANGELOG=internal