memtx: accelerate select with offset in tree
This patch rewrites the memtx tree iterator creation path so it handles
the offset parameter taking advantage of offset-based functions of the
BPS tree with BPS_INNER_CARD config.

It also introduces the memtx_tx_index_invisible_count_matching_until
function which counts the amount of tuples matching with a key and
iterator type until a tuple that are invisible to a transaction.

Since the invisible count is required in order to update the tree
iterator, the garbage collection is not performed in it anymore, it
is performed in the `index_size` methods of memtx indexes instead.

Part of #8204

NO_DOC=perf improvement