datetime: test RFC 3339 and ISO 8601 formats
The commit 3c40366172e3 ("datetime, lua: date parsing functions")
introduced a function `datetime.parse()`. The function accepts
a datetime format: 'iso8601', 'rfc3339' and a custom datetime
format. However, 'iso8601' as well as 'rfc3339' is a set of
datetime formats (see [1] and [2]), and `datetime.parse()`
supports a subset of these formats. There is a documentation
issue [3], where we should describe what formats are supported
and what formats are not. Supported formats must be added to
the testing and the patch do it.

1. https://datatracker.ietf.org/doc/html/rfc3339
2. https://ijmacd.github.io/rfc3339-iso8601/
3. https://github.com/tarantool/doc/issues/3155

Needed for tarantool/doc#3155

NO_CHANGELOG=testing
NO_DOC=gh-3155