datetime: test RFC 3339 and ISO 8601 formats
The commit 3c40366172e3 ("datetime, lua: date parsing functions")
introduced a function `datetime.parse()`. The function accepts
a datetime format: 'iso8601', 'rfc3339' and a custom datetime
format. However, 'iso8601' as well as 'rfc3339' is a set of
datetime formats (see [1] and [2]), and `datetime.parse()`
supports a subset of these formats. There is a documentation
issue [3], where we should describe what formats are supported
and what formats are not.

The patch adds a tests for most of supported formats.
The tests were created using a page about date and time formats
in ISO 8601 and RFC 3339 standards [2] and these tests are not
exhaustive. More thorough testing can be done with property-based
testing and randomization. It will be a subject of the following
commits.

1. https://datatracker.ietf.org/doc/html/rfc3339
2. https://ijmacd.github.io/rfc3339-iso8601/
3. https://github.com/tarantool/doc/issues/3155

Follows up #6731
Needed for tarantool/doc#3155

Co-authored-by: Sergey Kaplun <skaplun@tarantool.org>

NO_CHANGELOG=testing
NO_DOC=tarantool/doc#3155