schema: introduce the `offset` param of `index:pairs` method
The purpose of the argument is identical to the one in the `select`:
it makes the iterator skip first N tuples prior to return.

The `box_index_iterator_after` is removed as unused, since it's only
been used by the Lua site for the `pairs` iterator creation.

Part of #8204

@TarantoolBot document
Title: A new `offset` parameter of the `index:pairs` method
Product: Tarantool
Since: 3.3.0

Now the `index:pairs` method supports the `offset` parameter which may
be used to skip the specified amount of tuples of the iterator. If the
skip caused the iterator to exhaust, it behaves as expected (no tuple
is returned).