box: introduce the create_iterator_with_offset read view method
The method may be used to implement the accelerated select with offset
in indexes that support it. The generic function creates a read view
iterator and then skips the required amount of tuples.

Since the tree read view is going to support the accelerated version,
created a stub in the memtx tree implementation throwing an error. The
stub is to be implemented in EE.

Part of tarantool/tarantool-ee#858

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal