box: introduce the create_iterator_with_offset read view method
The method may be used to implement the accelerated select with offset
in indexes that support it. The function creates a regular read view
iterator and then skips the required amount of tuples if read views
are supported. Otherwise it throws an error.

Part of tarantool/tarantool-ee#858

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal