box: introduce the index read view count method
The method is meant to be used in order to count the amount of tuples
matching a key in the index read view. The generic version creates an
iterator and simply counts the tuples returned if read view supported.
Throws an error otherwise.

The memtx tree index overrides the method, but the implementation is
meant to be unreachable, the actual functionality is to be implemented
in the Tarantool EE.

Part of tarantool/tarantool-ee#858

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal