box: don't require functional index function to be sandboxed
Closes #4765
Closes #10349

@TarantoolBot document
Title: Functional index function doesn't have to sandboxed anymore

The user may now create a functional index using a function that doesn't
have the flag `is_sandboxed`. Note that a functional index function must
be deterministic and isn't allowed to yield or access the database so
the flag `is_sandboxed` is still recommended. However, there are cases
when the user might want to clear the flag. For example, it's the only
way to let a functional index function to call a deterministic function
defined in a module (e.g. `json.encode`).