memtx: make the memtx_tree_lookup method view-compatible
The function is used in order to perform a lookup into a tree by a
key and other parameters, but does not support search in a tree view.
Fix this by introducing a new template parameter and creating wrappers
for regular tree lookup functions.

NO_DOC=internal
NO_TEST=to be tested in EE
NO_CHANGELOG=internal