limbo: add event for asynchronously committed txn rollback
Add a `box.ctl.on_replication_split_brain_rollback` event for notification
of users about the rollback of asynchronously committed synchronous
transactions. This rollback can happen only because of a PROMOTE request
from a different node.

The event is delivered before writing a PROMOTE request for each
asynchronously committed transaction in the limbo that will be rolled back
by the PROMOTE request. This ensures that if the event trigger fails, the
PROMOTE request will also fail, ER_SPLIT_BRAIN will be raised, and the
asynchronously committed data will not be lost. The trigger is not called
during recovery of a PROMOTE request.

The event trigger receives a transaction statement iterator similar to
that of other transaction event triggers. The iterator yields all
transaction statement information as it was encoded to the WAL. The
information is decoded from `stmt→row` and pushed onto the Lua stack in a
format similar to the `xlog` module.

The trigger allows to execute a fully local transaction bypassing the
limbo, which is started before the trigger execution. The transaction is
committed implicitly after the trigger execution finishes if it is not
committed explicitly by the user. Failure to implicitly commit the
transaction also raises ER_SPLIT_BRAIN. Trying to run another transaction
other than the provided one (i.e., by explicitly beginning a new
transaction) will always fail.

Closes #10460

@TarantoolBot document
Title: `box.ctl.on_replication_split_brain_rollback` event
Product: Tarantool
Since: 3.4
Root document: https://www.tarantool.io/en/doc/latest/reference/reference_lua/box_txn_management/commit/
and https://www.tarantool.io/ru/doc/latest/reference/reference_lua/box_txn_management/atomic/
and https://www.tarantool.io/en/doc/latest/platform/replication/repl_sync/

Added a new `box.ctl.on_replication_split_brain_rollback` event for
notification of users about the rollback of asynchronously committed
synchronous transactions. This rollback can happen only because of a
PROMOTE request from a different node.

The event is delivered before writing a PROMOTE request for each
asynchronously committed transactions in the limbo that will be rolled back
by the PROMOTE request. This ensures that if the event trigger fails, the
PROMOTE request will also fail, ER_SPLIT_BRAIN will be raised, and the
asynchronously committed data will not be lost. The trigger is not called
during recovery of a PROMOTE request.

The event trigger receives a transaction statement iterator similar to
that of other transaction event triggers. The iterator yields all
transaction statement information as it was encoded to the WAL. The
information is decoded from `stmt→row` and pushed onto the Lua stack in a
format similar to the `xlog` module. The iterator returns MsgPack objects
with the following structure:
```lua
 {
    meta = {
        replica_id = ...,
        lsn = ...,
        term = ...,
    },
    statement = {
        op_type = ...,
        space_id = ...,
        key = ...,
        tuple = ...,
        ops = ...,
        old_tuple = ...,
        new_tuple = ...,
    }
}
```
The values of the `key`, `tuple`, and `ops` fields depend on the operation
type.

The semantics of the event require its trigger to be called multiple times
if the PROMOTE request or the trigger fail. Therefore, the trigger must be
idempotent with respect to transactions. The `lsn` value that is yielded by
the transaction statement iterator is sufficient to uniquely determine
transaction statements.

The trigger allows to execute a fully local transaction bypassing the
limbo, which is started before the trigger execution. The transaction is
committed implicitly after the trigger execution finishes if it is not
committed explicitly by the user. Failure to implicitly commit the
transaction also raises ER_SPLIT_BRAIN. Trying to run another transaction
other than the provided one (i.e., by explicitly beginning a new
transaction) will always fail.