limbo: add event for asynchronously committed synchronous txn rollback
Add a `box.ctl.on_replication_split_brain_rollback` event for notification
of users about the rollback of asynchronously committed synchronous
transactions. This rollback can happen only because of a PROMOTE request
from a different node.

The event is delivered before writing a PROMOTE request for each
asynchronous committed synchronous transaction in the limbo that will be
rolled back by the PROMOTE request. This ensures that if the event trigger
fails, the PROMOTE request will also fail, ER_SPLIT_BRAIN will be raised,
and the asynchronously committed data will not be lost. The trigger is not
called during recovery of a PROMOTE request.

The event trigger receives a transaction statement iterator similar to
that of other transaction event trigger. The iterator yields all
transaction statement information as it was encoded to the WAL. The
information is decoded from `stmt→row` and pushed onto the Lua stack
in the same format as for the `xlog` module.

The trigger allows to run fully local transactions, which is controlled by
a new limbo flag. This flag makes fully local transactions bypass the
limbo. If there is an active transaction after the execution of the event
triggers is finished, it is implicitly committed, and the commit failure
also raises ER_SPLIT_BRAIN.

Closes #10460

@TarantoolBot document
Title: `box.ctl.on_replication_split_brain_rollback` event
Product: Tarantool
Since: 3.4
Root document: https://www.tarantool.io/en/doc/latest/reference/reference_lua/box_txn_management/commit/
and https://www.tarantool.io/ru/doc/latest/reference/reference_lua/box_txn_management/atomic/
and https://www.tarantool.io/en/doc/latest/platform/replication/repl_sync/

Added a new `box.ctl.on_replication_split_brain_rollback` event for
notification of users about the rollback of asynchronously committed
synchronous transactions. This rollback can happen only because of a
PROMOTE request from a different node.

The event is delivered before writing a PROMOTE request for each
asynchronous committed synchronous transaction in the limbo that will be
rolled back by the PROMOTE request. This ensures that if the event trigger
fails, the PROMOTE request will also fail, ER_SPLIT_BRAIN will be raised,
and the asynchronously committed data will not be lost. The trigger is not
called during recovery of a PROMOTE request.

The event trigger receives a transaction statement iterator similar to
that of other transaction event trigger. The iterator yields all
transaction statement information as it was encoded to the WAL. The
information is pushed onto the Lua stack in the same format as for the
`xlog` module.

The semantics of the event require its trigger to be called multiple times
if the PROMOTE request or the trigger fail. Therefore, the trigger must be
idempotent with respect to transactions. The `lsn` value that is yielded by
the transaction statement iterator is sufficient to uniquely determine
transaction statements.

The trigger allows to run fully local transactions. If there is an active
transaction after the execution of the event triggers is finished, it is
implicitly committed, and the commit failure also raises ER_SPLIT_BRAIN.