box: introduce `TXN_IS_SYNC` flag
In the scope of #10459 we are going to introduce a new type of
transactions: asynchronous transactions that are immediately acked to the
user, but kept in the limbo until they get confirmed, just like synchronous
transactions.

Currently, transactions kept in the limbo are identified by the
`TXN_WAIT_SYNC` flag. However, this flag means that a transaction is
blocked until confirmation and is not immediately acked to the user.

Apparently, the new type of transactions described above will not have this
flag set (unless they are blocked by preceding synchronous transactions),
so in order to identify transactions kept in the limbo, let's introduce a
new internal `TXN_IS_SYNC` flag.

Needed for #10459

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>