ci: add jobs for testing LuaJIT allocator & ASAN
This patch is a pre-patch (refer to [1]) to add ASAN instrumentation to
the LuaJIT memory allocator.

LuaJIT has two usage scenarios for ASAN:
- LuaJIT using sysmalloc
- LuaJIT using internal memory allocator

This patch introduces jobs that separate the testing of these scenarios
into individual jobs. This is necessary because using the embedded
allocator with ASAN requires `-DLUAJIT_ENABLE_GC64=ON` (refer to [1]).

Note: this patch is required to test the ASAN instrumentation of LuaJIT
internal memory allocator.

[1]: Issue #10231

NO_DOC=ci
NO_CHANGELOG=ci
NO_TEST=ci