box: fix several corner cases of deletion of replica on deleted replica
Firstly, there was a bug in the test: since we do not use a full-mesh
topology in the `g_three_member_cluster` test group (and hence we cannot
use `wait_for_fullmesh`), we need to explicitly wait for the `replica` and
the `to_be_deleted` nodes to synchronize.

Secondly, we need to set the `FiberIsCancelled` diagnostic for the new
check after acquisition of the applier latch.

Finally, we need to reset the relay's subscribe fiber, when it gets stopped
on its own, in order to avoid double cancel.

The fixes are hard to test, since they require synchronization of applier
and relay states during the tests, so let's rely on the existing test,
which detected these bugs.

Follows up #10088
Closes #10397

NO_CHANGELOG=<change not released>
NO_DOC=<bugfix>