failover: revoke leader RW in supervised mode
The bootstrap leader has to make itself RW right after bootstrapping to
let all the triggers proceed. It causes some problems for the supervised
failover because right now the failover assumes it's the only one
instance allowed to manage other replica modes.

The patch makes the bootstrap leader automatically go RO after the
`lease_intrval` after bootstrapping. For this it uses a separate fiber
in case the `replication.failover` is `supervised`.

Also, the patch adds the `_recently_bootstraped` variable to the config
singleton in order to be able to detect the possibility of the instance
to go to RW mode without the failover involvement. The variable is set
for the `lease_interval` after completing the bootstrap process, i.e
starting up without any snapshot.

Closes #10419