box: account for `wait_ack` flag of the tx's last row during recovery
Currently, we do not account for the `wait_ack` flag of the transaction's
last row during recovery, instead relying on the fact the transaction will
be made synchronous when applying operations on a synchronous space.

However, this logic is at odds with explicitly created synchronous
transactions (created with box.begin{is_sync = true}), which may not have
operations on synchronous spaces. As a result, during recovery we do not
treat these transactions as synchronous.

Let's fix this by mirroring the applier logic to the recovery.

Closes #10412

NO_DOC=<bugfix>