box: refactor writing and reading confirm into a separate helper
Writing a confirm always entails reading it, so let's refactor it into a
separate helper.

The patch is cherry-picked to 2.11 in order to bring the limbo's
state as close to the master (3.6) as possible in order to
fix #11790.

Needed for #10628
In scope of #11790

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>

(cherry picked from commit bc057ed71b7e98f1749019302ae25b512eb597a5)