test: fix flaky gh_9918_synchro_queue_additional_info_test.lua
This patch eliminates flaking in the
`gh_9918_synchro_queue_additional_info_test.lua` test. The problem was
that the test did not wait for the connection between master and replica
to be established, and therefore master node was in the "orphan" state at
the time `box.ctl.promote()` was called. Thus, it turned out that the
master node became the owner of the limbo, but was still read only.

To fix this, this patch simply calls `cg.replica_set:wait_for_fullmesh()`
on the previous line before `box.ctl.promote()`.

Closes #10463

NO_DOC=test fix
NO_CHANGELOG=test fix