memtx: refactor the nearby gap key saving out
The nearby gap item requires a key to be used, the key is either
allocated on the transaction region or in the gap item itself. This
logic will be used when count gap item is updated to fit the select
with offset usecase, so let's refactor it into a separated function.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring